# packages/client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "operationId": "Api_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "login",
                  "password",
                  "nickname"
                ],
                "properties": {
                  "login": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "nickname": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tokens"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [],
        "summary": "Регистрация",
        "tags": [
          "Focus Coins"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "Api_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "login",
                  "password"
                ],
                "properties": {
                  "login": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tokens"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [],
        "summary": "Вход",
        "tags": [
          "Focus Coins"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "Api_refresh",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tokens"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [],
        "tags": [
          "Focus Coins"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                },
                "required": []
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "Api_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [],
        "tags": [
          "Focus Coins"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                },
                "required": []
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard": {
      "get": {
        "operationId": "Api_dashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Focus Coins"
        ]
      }
    },
    "/api/v1/leaderboard": {
      "get": {
        "operationId": "Api_leaderboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Leaderboard"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Focus Coins"
        ]
      }
    },
    "/api/v1/focus": {
      "post": {
        "operationId": "Api_start",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "minutes"
                ],
                "properties": {
                  "minutes": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 360
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Focus Coins"
        ]
      }
    },
    "/api/v1/focus/{id}/cancel": {
      "post": {
        "operationId": "Api_cancel",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Focus Coins"
        ]
      }
    },
    "/api/v1/games/{kind}": {
      "post": {
        "operationId": "Api_play",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "stake"
                ],
                "properties": {
                  "stake": {
                    "type": "string",
                    "pattern": "^\\d+(\\.\\d{1,2})?$"
                  },
                  "bet": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": [
                          "red",
                          "black",
                          "even",
                          "odd",
                          "number"
                        ]
                      },
                      "number": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 36
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Round"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Focus Coins"
        ]
      }
    },
    "/api/v1/tower/{id}/pick": {
      "post": {
        "operationId": "Api_pick",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "cell"
                ],
                "properties": {
                  "cell": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Round"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Focus Coins"
        ]
      }
    },
    "/api/v1/tower/{id}/collect": {
      "post": {
        "operationId": "Api_collect",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Round"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Focus Coins"
        ]
      }
    },
    "/api/v1/profile": {
      "patch": {
        "operationId": "Api_profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nickname"
                ],
                "properties": {
                  "nickname": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Focus Coins"
        ]
      }
    },
    "/api/v1/profile/password": {
      "post": {
        "operationId": "Api_changePassword",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Focus Coins"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "current": {
                    "type": "string"
                  },
                  "next": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 128
                  }
                },
                "required": [
                  "current",
                  "next"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/visits": {
      "post": {
        "operationId": "Api_visit",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Conflicting state or idempotency key"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Focus Coins"
        ]
      }
    },
...
  },
  "info": {
    "title": "Focus Coins API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "minutes": {
            "type": "integer"
          },
          "startedAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "reward": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d{2}$",
            "example": "100.00"
          },
          "multiplier": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "minutes",
          "startedAt",
          "endsAt",
          "status",
          "reward",
          "multiplier"
        ]
      },
      "Round": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "stake": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d{2}$",
            "example": "100.00"
          },
          "payout": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d{2}$",
            "example": "100.00"
          },
          "net": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d{2}$",
            "example": "100.00"
          },
          "status": {
            "type": "string"
          },
          "result": {
            "type": "object",
            "additionalProperties": true
          },
          "floors": {
            "type": "integer"
          },
          "picks": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "expiresAt": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "stake",
          "payout",
          "net",
          "status",
          "result",
          "floors",
          "picks",
          "expiresAt",
          "createdAt"
        ]
      },
      "Dashboard": {
        "type": "object",
        "properties": {
          "serverTime": {
            "type": "string"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "login": {
                "type": "string"
              },
              "nickname": {
                "type": "string"
              },
              "createdAt": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "login",
              "nickname",
              "createdAt"
            ]
          },
          "season": {
            "type": "object",
            "properties": {
              "key": {
                "type": "string"
              },
              "endsAt": {
                "type": "string"
              }
            },
            "required": [
              "key",
              "endsAt"
            ]
          },
          "balance": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d{2}$",
            "example": "100.00"
          },
          "reserved": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d{2}$",
            "example": "100.00"
          },
          "rankingBalance": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d{2}$",
            "example": "100.00"
          },
          "rank": {
            "type": "integer"
          },
          "active": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Session"
              }
            ]
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Session"
            }
          },
          "rounds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Round"
            }
          },
          "tower": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Round"
              }
            ]
          },
          "daily": {
            "type": "object",
            "properties": {
              "minutes": {
                "type": "integer"
              },
              "streak": {
                "type": "integer"
              },
              "cycle": {
                "type": "integer"
              },
              "awarded": {
                "type": "boolean"
              },
              "boostUntil": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "minutes",
              "streak",
              "cycle",
              "awarded",
              "boostUntil"
            ]
          },
          "stats": {
            "type": "object",
            "properties": {
              "totalMinutes": {
                "type": "integer"
              },
              "weekMinutes": {
                "type": "integer"
              },
              "monthMinutes": {
                "type": "integer"
              },
              "completed": {
                "type": "integer"
              },
              "cancelled": {
                "type": "integer"
              },
              "focusCoins": {
                "type": "string",
                "pattern": "^-?\\d+\\.\\d{2}$",
                "example": "100.00"
              },
              "bonusCoins": {
                "type": "string",
                "pattern": "^-?\\d+\\.\\d{2}$",
                "example": "100.00"
              },
              "gameNet": {
                "type": "string",
                "pattern": "^-?\\d+\\.\\d{2}$",
                "example": "100.00"
              },
              "days": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "day": {
                      "type": "string"
                    },
                    "minutes": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "day",
                    "minutes"
                  ]
                }
              }
            },
            "required": [
              "totalMinutes",
              "weekMinutes",
              "monthMinutes",
              "completed",
              "cancelled",
              "focusCoins",
              "bonusCoins",
              "gameNet",
              "days"
            ]
          },
          "ledger": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "kind": {
                  "type": "string"
                },
                "amount": {
                  "type": "string",
                  "pattern": "^-?\\d+\\.\\d{2}$",
                  "example": "100.00"
                },
                "createdAt": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "kind",
                "amount",
                "createdAt"
              ]
            }
          }
        },
        "required": [
          "serverTime",
          "user",
          "season",
          "balance",
          "reserved",
          "rankingBalance",
          "rank",
          "active",
          "history",
          "rounds",
          "tower",
          "daily",
          "stats",
          "ledger"
        ]
      },
      "Rank": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "balance": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d{2}$",
            "example": "100.00"
          },
          "rank": {
            "type": "integer"
          }
        },
        "required": [
          "userId",
          "nickname",
          "balance",
          "rank"
        ]
      },
      "Leaderboard": {
        "type": "object",
        "properties": {
          "season": {
            "type": "string"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rank"
            }
          },
          "self": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Rank"
              }
            ]
          },
          "archives": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "season",
          "rows",
          "self",
          "archives"
        ]
      },
      "Tokens": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      }
    }
  }
}